Update patch set 3
Patch Set 3:

Just a heads up since I see that you are copying the docs layout from ONAP. We're developing a new Python package [0] which will allow us to share conf.py configuration across projects without needed to load all the projects as a submodule inside of a master docs project and allowing all projects to push their docs to their own RTD projects. (Sharing conf.py was the reason ODL and OPNFV started out with submodule projects but we're finding that as docs get larger and larger build times increase tremendously and maintenance of submodule becomes more difficult).

This workflow is NOT ready yet but I'll keep in touch when it is and we can discuss if Acumos would be interested in following it.

[0] https://gerrit.linuxfoundation.org/infra/7463

Patch-set: 3