Add missing validation-security submodule
The documentation-rtd-verify-master job is failing because
docs/DevelopingAcumos/DevelopmentGuides.rst has an entry for
a validation-security submodule directory but that directory does
not exist.

Git does not allow committing an empty directory. I added an RST
file to trigger a build. After success, I deleted the RST file and
added a .keep so the directory structure remains intact.

Change-Id: Ib93166f233889d3b7153bc725eef640a77f54d16
Signed-off-by: Aimee Ukasick <aimeeu.opensource@gmail.com>