Update patch set 1
Patch Set 1:

The script does not automatically create a log. But I ave run this dozens of times successfully, and this patch anyway is to the docs, not the script. So I don't think you need to test the script to merge this patch.

Patch-set: 1
Reviewer: Gerrit User 93 <93@655300a2-7a56-4504-9153-881edd907bea>
Label: Code-Review=0