Update patch set 3
Patch Set 3: Code-Review+1

For the moment I have no idea when we will disable the old method with username and password. It will be a period during which the two methods will coexist then we will broke the old method. So I think for the moment the best way is to explain the two methods in the documentation and then we will update it once the old method will have been deleted.

Patch-set: 3
Reviewer: Gerrit User 212 <212@655300a2-7a56-4504-9153-881edd907bea>
Label: Code-Review=+1