Update patch set 1
Patch Set 1:

Merged text removed. Git adds "stuff" sometimes and I don't always catch it. I think this happened when you merged the last patch and I didn't see that, and was going to push this update. I got the merge conflict notice and just recloned to start over. I guess it was halfway and had added the stuff to the file.

Patch-set: 1
Label: Code-Review=0