Update patch set 1
Patch Set 1: Code-Review-1

The is this line refering to the top search or the search on the models page?"
  99 #. Enter keywords in the search field on the top left inner menu" 

Only one search is identified clearly.  There are two searches, one as a global and one as a search that includes filters on the models page on the left.  Which is being identified in the documentation.

Patch-set: 1
Label: Code-Review=-1