Update patch set 1
Patch Set 1: Code-Review-1

AcumosContributor/component-guides.html needs to be edited; line   47 * :doc:`License Manager Client Library<../submodules/security-verification/license-manager-client-library/docs/index>`   ( these links use to be in /security-verification/license-manager-client-library) needs to be updated to point to the new repo. A link needs to be added for are the license-usage-manager.

Patch-set: 1
Label: Code-Review=-1