Update patch set 1
Patch Set 1:

(1 comment)

> Patch Set 1:
> 
> (1 comment)
> 
> what build issue? without a new/changed rst file, the documentation jenkins job does not build

Hi Lorraine,

The build issue is related to moving license-manager docs from security-verfication repo to it's own repo in Clio. We are not seeing the docs for license-manager after we moved it. Also we are including license-usage-manager repo for the LUM component that is being added in Clio.

Thanks,

Justin

Patch-set: 1
Label: Code-Review=0