Update patch set 1
Patch Set 1: Code-Review+1

It looks to me like the verify job doesn't actually test changes to the .gitmodules. If you look at the verify job for this change, it is not pulling in the new submodule. And the merge job just triggers the RTD build, so that doesn't test it either.

So I think what we need to do (besides this revert, which absolutely needs to happen) is to fix the documentation verify job, so that it actually will pull in changes to the .gitmodules file.

Patch-set: 1
Reviewer: Gerrit User 51 <51@655300a2-7a56-4504-9153-881edd907bea>
Reviewer: Gerrit User 223 <223@655300a2-7a56-4504-9153-881edd907bea>
Label: Code-Review=+1