Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review+1
> 
> It looks to me like the verify job doesn't actually test changes to the .gitmodules. If you look at the verify job for this change, it is not pulling in the new submodule. And the merge job just triggers the RTD build, so that doesn't test it either.
> 
> So I think what we need to do (besides this revert, which absolutely needs to happen) is to fix the documentation verify job, so that it actually will pull in changes to the .gitmodules file.

Sorry, when I said "the verify job for this change" above, I meant the original change which is being reverted (https://gerrit.acumos.org/r/c/documentation/+/7986). But similarly, this verify job is failing because it is not using the updated .gitmodules that this change creates; it's using the one from master.

Patch-set: 1
Reviewer: Gerrit User 223 <223@655300a2-7a56-4504-9153-881edd907bea>